Updated module augmentation for block-quote package.
