+++ packages/ckeditor5-block-quote/src/blockquotecommand.ts
@@ -220,9 +220,3 @@ function checkCanBeQuoted( schema: Schema, block: Element ): boolean {
 
 	return isBQAllowed && isBlockAllowedInBQ;
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		blockQuote: BlockQuoteCommand;
-	}
-}
