+++ packages/ckeditor5-block-quote/src/index.ts
@@ -10,3 +10,5 @@
 export { default as BlockQuote } from './blockquote';
 export { default as BlockQuoteEditing } from './blockquoteediting';
 export { default as BlockQuoteUI } from './blockquoteui';
+
+import './augmentation';
