+++ packages/ckeditor5-build-classic/src/ckeditor.ts
@@ -10,7 +10,7 @@ import Essentials from '@ckeditor/ckeditor5-essentials/src/essentials';
 import { UploadAdapter } from '@ckeditor/ckeditor5-adapter-ckfinder';
 import { Autoformat } from '@ckeditor/ckeditor5-autoformat';
 import { Bold, Italic } from '@ckeditor/ckeditor5-basic-styles';
-import BlockQuote from '@ckeditor/ckeditor5-block-quote/src/blockquote';
+import { BlockQuote } from '@ckeditor/ckeditor5-block-quote';
 import CKBox from '@ckeditor/ckeditor5-ckbox/src/ckbox';
 import CKFinder from '@ckeditor/ckeditor5-ckfinder/src/ckfinder';
 import EasyImage from '@ckeditor/ckeditor5-easy-image/src/easyimage';
